+++ src/server/Worker.ts
@@ -7,6 +7,10 @@ import path from "path";
 import { fileURLToPath } from "url";
 import { WebSocket, WebSocketServer } from "ws";
 import { z } from "zod";
+import {
+  ClanExistsResponseSchema,
+  clanExistsApiPath,
+} from "../core/ApiSchemas";
 import { GameEnv } from "../core/configuration/Config";
 import { GameType } from "../core/game/Game";
 import {
@@ -21,7 +25,7 @@ import { archive, finalizeGameRecord } from "./Archive";
 import { Client } from "./Client";
 import { GameManager } from "./GameManager";
 import { registerGamePreviewRoute } from "./GamePreviewRoute";
-import { clanExistsByTag, getUserMe, verifyClientToken } from "./jwt";
+import { getUserMe, verifyClientToken } from "./jwt";
 import { logger } from "./Logger";
 
 import { MapPlaylist } from "./MapPlaylist";
@@ -37,6 +41,74 @@ import { initWorkerMetrics } from "./WorkerMetrics";
 const workerId = ServerEnv.workerId() ?? 0;
 const log = logger.child({ comp: `w_${workerId}` });
 const playlist = new MapPlaylist();
+
+// Clan-existence probe used by the join-time ownership check.
+// Caches results briefly so a lobby surge doesn't fan out to the auth API.
+// Returns null on transport errors / unexpected statuses so callers fail open.
+const CLAN_EXISTS_FETCH_TIMEOUT_MS = 3000;
+const CLAN_EXISTS_CACHE_TTL_MS = 60_000;
+const clanExistsCache = new Map<
+  string,
+  { result: boolean; expiresAt: number }
+>();
+
+async function clanExistsByTag(tag: string): Promise<boolean | null> {
+  const cacheKey = tag.toUpperCase();
+  const entry = clanExistsCache.get(cacheKey);
+  if (entry !== undefined) {
+    if (Date.now() < entry.expiresAt) return entry.result;
+    clanExistsCache.delete(cacheKey);
+  }
+
+  try {
+    const url = `${ServerEnv.jwtIssuer()}${clanExistsApiPath(tag)}`;
+    const response = await fetch(url, {
+      headers: { Accept: "application/json" },
+      signal: AbortSignal.timeout(CLAN_EXISTS_FETCH_TIMEOUT_MS),
+    });
+    if (response.status === 200) {
+      // Upstream currently has no body; tolerate {exists:false} for forward-compat.
+      try {
+        const text = await response.text();
+        if (text.length > 0) {
+          const parsed = ClanExistsResponseSchema.safeParse(JSON.parse(text));
+          if (parsed.success && parsed.data?.exists === false) {
+            clanExistsCache.set(cacheKey, {
+              result: false,
+              expiresAt: Date.now() + CLAN_EXISTS_CACHE_TTL_MS,
+            });
+            return false;
+          }
+        }
+      } catch {
+        // Forward-compat parsing only; ignore failures.
+      }
+      clanExistsCache.set(cacheKey, {
+        result: true,
+        expiresAt: Date.now() + CLAN_EXISTS_CACHE_TTL_MS,
+      });
+      return true;
+    }
+    if (response.status === 404) {
+      clanExistsCache.set(cacheKey, {
+        result: false,
+        expiresAt: Date.now() + CLAN_EXISTS_CACHE_TTL_MS,
+      });
+      return false;
+    }
+    log.warn("clanExistsByTag: unexpected status, failing open", {
+      tag: cacheKey,
+      status: response.status,
+    });
+    return null;
+  } catch (e) {
+    log.warn("clanExistsByTag: fetch failed, failing open", {
+      tag: cacheKey,
+      error: e instanceof Error ? e.message : String(e),
+    });
+    return null;
+  }
+}
 
 // Worker setup
 export async function startWorker() {
